Work around weird Firefox behavior
Firefox recently started resolving the decode() promise
immediately if an image had been decoded in the past. This
happens even though the line above changes src. To work around it
we first clear src and then set it. We use setAttribute because
doing it with .src doesn't seem to work around it.